[BUGFIX] Translating a contained element
When a single element in a container gets translated (from within
the Translation Drop-Down in FormEngine / while editing the element)
the pointer to the parent container does not get updated.

So the translated content element still points to the untranslated
parent container and will not get shown in Page-View.

This patch fixes this issue: Wtranslating a single element it has to
determine its parent container, find a translation of it and assign
the translated container to the translated element.

Resolves: #65380
Releases: master



git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/kb_nescefe/trunk@89251 735d13b6-9817-0410-8766-e36946ffe9aa