[FEATURE] Added file select wizards to the template fields
When editing a kb_nescefe container layout record until now
the path to the template files had to get entered manually.
Altough the "EXT:" prefix could get used to specify a file
residing in an extension directory.

This features adds a file picker wizard to the FE and BE
template fields. Since FAL is used for referencing files the
resulting value in the FE/BE template field will be a value
like "file:123" where 123 designates the sys_file uid of the
file being used.

Resolves: #31521
Releases: master



git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/kb_nescefe/trunk@89256 735d13b6-9817-0410-8766-e36946ffe9aa